+++ src/main/java/zone/vao/nexoAddon/NexoAddon.java
@@ -43,9 +43,9 @@ import zone.vao.nexoAddon.events.nexo.furnitures.NexoFurnitureInteractListener;
 import zone.vao.nexoAddon.events.nexo.furnitures.NexoFurniturePlaceListener;
 import zone.vao.nexoAddon.events.player.PlayerInteractListener;
 import zone.vao.nexoAddon.events.player.PlayerMovementListener;
-import zone.vao.nexoAddon.handlers.BlockHardnessHandler;
-import zone.vao.nexoAddon.handlers.ParticleEffectManager;
-import zone.vao.nexoAddon.handlers.RecipeManager;
+import zone.vao.nexoAddon.utils.handlers.BlockHardnessHandler;
+import zone.vao.nexoAddon.utils.handlers.ParticleEffectManager;
+import zone.vao.nexoAddon.utils.handlers.RecipeManager;
 import zone.vao.nexoAddon.utils.hooks.PacketEventsHook;
 import zone.vao.nexoAddon.utils.metrics.Metrics;
 import zone.vao.nexoAddon.populators.CustomChunkGenerator;
