+++ src/main/java/zone/vao/nexoAddon/events/PlayerCommandPreprocessListener.java
@@ -4,7 +4,7 @@ import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerCommandPreprocessEvent;
-import zone.vao.nexoAddon.handlers.RecipeManager;
+import zone.vao.nexoAddon.utils.handlers.RecipeManager;
 import zone.vao.nexoAddon.utils.RecipesUtil;
 
 public class PlayerCommandPreprocessListener implements Listener {
