+++ src/main/java/zone/vao/nexoAddon/events/nexo/NexoPackUploadListener.java
@@ -5,7 +5,7 @@ import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.scheduler.BukkitRunnable;
 import zone.vao.nexoAddon.NexoAddon;
-import zone.vao.nexoAddon.handlers.RecipeManager;
+import zone.vao.nexoAddon.utils.handlers.RecipeManager;
 import zone.vao.nexoAddon.utils.RecipesUtil;
 
 public class NexoPackUploadListener implements Listener {
