+++ src/main/java/zone/vao/nexoAddon/utils/handlers/ApiCompatibilityHandler.java
@@ -1,4 +1,4 @@
-package zone.vao.nexoAddon.handlers;
+package zone.vao.nexoAddon.utils.handlers;
 
 import com.nexomc.nexo.mechanics.furniture.FurnitureMechanic;
