+++ src/main/java/zone/vao/nexoAddon/utils/handlers/ParticleEffectManager.java
@@ -1,4 +1,4 @@
-package zone.vao.nexoAddon.handlers;
+package zone.vao.nexoAddon.utils.handlers;
 
 import com.nexomc.nexo.api.NexoItems;
 import net.objecthunter.exp4j.ExpressionBuilder;
